documentation_agent/requirements.txt
claude-code-sdk>=0.0.20
python-dotenv>=1.0.0
boto3>=1.34.0
awscli>=1.32.0